+++ Assets/Imports/Mirror/Runtime/Transport/EpicOnlineTransport/EOSSDKComponent.cs
@@ -9,7 +9,6 @@ using UnityEngine;
 using UnityEngine.Events;
 
 using CandyCoded.env;
-using UnityEngine.Events;
 
 /// <summary>
 /// Manages the Epic Online Services SDK
@@ -393,13 +392,12 @@ namespace EpicTransport {
                 Result result = loginCallbackInfo.LocalUserId.ToString(out productIdString);
                 if (Result.Success == result) {
                     Debug.Log("EOS User Product ID:" + productIdString);
-                    OnConnected.Invoke();
 
                     localUserProductIdString = productIdString;
                     localUserProductId = loginCallbackInfo.LocalUserId;
 
                     Guid.TryParse(productIdString, out Guid guid);
-                    OnConnected.Invoke(guid);
+                    OnConnected.Invoke();
                 }
 
                 initialized = true;
