+++ Assets/Scripts/ConnectionsManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a2d73feaabfe9c54ba1d9da691e2b391
+guid: d0f70232b7995844ebf697748c21e9ff
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
