+++ Assets/Scripts/ServerConnection.cs
@@ -1,8 +1,6 @@
 using System;
 using System.Net;
-using System.Threading;
 using System.Threading.Tasks;
-using System.Collections.Generic;
 using UnityEngine;
 
 using SuperSimpleTcp;
